+++ tests/testthat/TestCompiled/R/TestCompiled.R
@@ -4,3 +4,7 @@
 simple <- function(x) {
   .Call(simple_, x) # nolint
 }
+
+simple2 <- function(x) {
+  .Call(simple2_, x) # nolint
+}